WEBNEW-102 :recycle: convert helpers and EventListingCard to typescript
