+++ package.json
@@ -77,6 +77,7 @@
     "@testing-library/react-hooks": "^3.2.1",
     "@types/jest": "^25.1.0",
     "@types/jest-axe": "^3.2.1",
+    "@types/moment-timezone": "^0.5.13",
     "@types/react": "^16.9.19",
     "@types/react-dates": "^17.1.7",
     "@types/react-dom": "^16.9.5",
