+++ src/events/index.ts
@@ -1,3 +1,3 @@
 export { default as EventsPage } from './EventsPage'
 export { default as EventPage } from './event/EventPage'
-export { default as EventCard } from './EventListingCard'
+export { default as EventCard } from './eventListingCard/EventListingCard'
